EGS_Mesh: combine all tests into test_egs_mesh.cpp
