--- HEN_HOUSE/egs++/geometry/egs_mesh/tests/five-tet.msh
@@ -1,40 +1,0 @@
-$MeshFormat
-4.1 0 8
-$EndMeshFormat
-$Entities
-0 0 0 1
-1 0 0 0 0 0 0 1 1 0
-$EndEntities
-$PhysicalNames
-1
-3 1 "Water"
-$EndPhysicalNames
-$Nodes
-1 8 1 8
-3 1 0 8
-1
-2
-3
-4
-5
-6
-7
-8
-0 0 0
-1 0 0
-0 1 0
-0 0 1
-0 -1 0
--1 0 0
-0 0 -1
-1 1 1
-$EndNodes
-$Elements
-1 5 1 5
-3 1 4 5
-1 1 2 3 4
-2 1 2 4 5
-3 1 3 4 6
-4 1 2 3 7
-5 2 3 4 8
-$EndElements
